+++ clang/test/Driver/mloadtime-comment-vars.c
@@ -1,11 +1,12 @@
 // RUN: %clang -### -target powerpc-ibm-aix -mloadtime-comment-vars=sccsid,version %s 2>&1 | FileCheck %s
 // RUN: %clang -### -target x86_64-linux-gnu -mloadtime-comment-vars=sccsid,version %s 2>&1 | FileCheck %s --check-prefix=NONAIX
 
-// Verify the option is forwarded verbatim to cc1 on AIX.
+// Verify the option is forwarded verbatim to cc1 for a supported target.
 // CHECK: "-cc1"
 // CHECK-SAME: "-mloadtime-comment-vars=sccsid,version"
 
-// Verify a warning is emitted and the option is NOT forwarded on non-AIX targets.
+// Verify a warning is emitted and the option is NOT forwarded for an
+// unsupported target.
 // NONAIX: warning: ignoring '-mloadtime-comment-vars=sccsid,version' option as it is not currently supported for target 'x86_64-unknown-linux-gnu'
 // NONAIX: "-cc1"
 // NONAIX-NOT: "-mloadtime-comment-vars=sccsid,version"